Display server history in Modal
Adds a new component named ServerSelector that displays a given list and calls a callback function when one of the items is selected.
This component is displayed in a Modal to provide the "select previously used server" function, replacing the combobox hack that looked pretty in Ubuntu but probably horrible everywhere else.