Make saving of scripts to LocalStorage analogous to server urls
Instead of adding a prefixed value to LocalStorage for each script, all scripts are now wrapped into a savedScripts object that is serialized into JSON.

Instead of reading from and writing to LocalStorage directly all the time, this object is now read from LocalStorage once when the component is mounted and after that a watcher takes care that any changes to it are saved back to LocalStorage.

This commit also adds a deleteScript method that was previously not there and changes CSS to avoid a line break in the toolbar on some (read: my) screens.