Replace prompt for script name with comfy list in modal
The savedScriptNames property is computed from the savedScripts property. Vue takes care that it is updated when something changes.

If loadScript is called without parameters, the modal is shown, if a script name is given as parameter, the loading is performed (as before).

Known bug: Removing items from the list doesn't update the view in the modal