Replace hardcoded credentials with CredentialsForm modal
Introduces a new CredentialsForm component that allows input of username and password and passes these to a submitCallback when the user clicks OK, or calls a cancelCallback when the user clicks Cancel.

This commit also introduces a new "modalClosed" event. This is thrown by the Modal component when its `close` method is called. This way the CredentialsForm can always execute the cancelCallback, even if the user doesn't cancel via the form's own Cancel button but via clicking the X in the modal's top right corner.