Allow functions as data source for List component
The `items` prop is replaced with a `dataSource` prop, which can be an array as before, but may alternatively be a function that returns an array.
The template still uses an `items` property, which is now a computed property. Its computation is dead simple: if `dataSource` is a function, evaluate that function and return its result; if it's something else, simply pipe through the value.