Fix #12
Introduces a new event `removeWebService` that undos everything the corresponding `viewWebService` did.

I also changed `i` to `id` everyhwere to make it clearer that we're iterating over ids, not over numbers.

I also changed the parameter of `addLayerToMap` from `layer` to `id` so that everyone's forced to put their layer into `this.layer` to get it on the map and no layer reference is "lost".