Improve UI for empty datasets
New convention: If there's data (i.e. `data.length` > 0), `noDataMessage` should be set to `undefined`.

If data retrieval was successful, the table header is displayed. That explicitly includes the case when data was retrieved but the dataset is empty.

If the dataset is empty, `noDataMessage` is the empty string since the emptiness of the dataset is already visualized by the empty table.

I.e. `noDataMessage` is only displayed if there's no data due to an actual error.
(Worded the other way round: `noDataMessage` is not displayed if there's no data to be displayed because the successfully loaded dataset is empty.)