Adapt BackendPanel to v0.3.0 changes
Mainly renaming of methods and attributes. The central `openEO` object is gone, so everything has to be passed seperately as props. Discovering processes now returns all available infos, so no extra request necessary anymore when user wants to see details, instead the cached data is used.

The old `openEO.Visualization` was used by both the BackendPanel and the SourceEnvironment, so there's a little change there as well. Not 100% about this whole visualizations thing though...