Improve handling of servers without Auth
If the server doesn't support authentication, the `serverChanged` event was emitted immediately, without the delay that normally occurs (due to the user being prompted the Auth modal). This fired the discoverData function in the BackendPanel before the new Capabilities object was propagated there by Vue. So better use nextTick to wait until the next update cycle of Vue has finished.

Also, it's nicer not to `throw` an exception when there's no Auth, because that's not *really* an error and because it would cause the UI to display the "Sorry server is not responding" message which is not the truth.