Add tour for credentials stage of ConnectForm
Specifying the targets directly is a lot more versatile. With the previous method of always adding a class to the element that is named after a convention, this tour wouldn't be possible like this (because it's impossible to add custom classes to elements that are created by the Tabs component, and because it wouldn't be possible to dynamically select the input field that is currently visible).